Widgets: the resize grip is not a bauhaus widget -- give it its own file
dt_bauhaus_resize_handle_new() and its three event handlers sat in bauhaus.c, which is the
slider/combobox toolkit. The grip has nothing to do with either: it never touches a
dt_bauhaus_widget_t, it is a GtkEventBox that turns a pointer drag into a size and hands
that size to whoever owns the resized target. It was in bauhaus.c because bauhaus.c was
where GTK widgets went.

Moved verbatim to widgets/resize_handle.{c,h} and renamed to dtgtk_resize_handle_*, which
is what the four other things it resizes already call it in spirit -- the main panels
(window_manager.c), the filmstrip and the side panel (gtk.c) and the histogram scope. The
"bauhaus" in the old name claimed a relationship that does not exist.

It stays under widgets/: its only non-GTK dependencies are dt_widget_set_cursor() and
DT_PIXEL_APPLY_DPI, both from widget_settings.h, so it satisfies the directory's rule as
it stands.
