+++ src/widgets/bauhaus.h
@@ -167,8 +167,6 @@ typedef struct dt_bauhaus_widget_t DtBauhausWidget;
 typedef struct dt_bauhaus_widget_class_t DtBauhausWidgetClass;
 
 typedef void (*dt_bauhaus_quad_paint_f)(cairo_t *cr, gint x, gint y, gint w, gint h, gint flags, void *data);
-typedef int (*dt_bauhaus_resize_handle_get_size_f)(gpointer user_data);
-typedef int (*dt_bauhaus_resize_handle_resize_f)(int requested_size, gboolean finished, gpointer user_data);
 
 // our new widget and its private members, inheriting from drawing area:
 typedef struct dt_bauhaus_widget_t
@@ -327,28 +325,6 @@ void dt_bauhaus_widget_set_field(GtkWidget *w, gpointer field, dt_introspection_
 
 void dt_bauhaus_hide_popup(dt_bauhaus_t *bh);
 void dt_bauhaus_show_popup(GtkWidget *w);
-
-/**
- * @brief Create a themed handle widget driving one-dimensional resize gestures.
- *
- * @details The handle owns the GTK event bookkeeping: hover state, cursor, grab lifetime,
- * drawing and drag delta computation. The caller owns the resized target and keeps that
- * ownership visible through @p get_size and @p resize. During pointer motion @p resize receives
- * `finished == FALSE`; on button release it receives `finished == TRUE` so callers can persist
- * the final size without writing settings at every motion sample.
- *
- * @param invert When FALSE the target grows as the pointer moves in the positive axis direction
- * (down for vertical, right for horizontal) — the natural case for a handle sitting below/at the
- * right of its target. Set TRUE when the target grows in the opposite direction, e.g. a right
- * panel that grows as it is dragged left, or a bottom panel that grows as it is dragged up.
- *
- * The grip is meant to be added as an overlay child on the resized widget. It pins itself to the
- * correct edge (from @p orientation and @p invert) and tags itself with an edge CSS class
- * (.resize-handle-{top,bottom,left,right}); its thickness and centering live in the stylesheet.
- */
-GtkWidget *dt_bauhaus_resize_handle_new(GtkOrientation orientation, gboolean invert, const char *tooltip,
-                                        dt_bauhaus_resize_handle_get_size_f get_size,
-                                        dt_bauhaus_resize_handle_resize_f resize, gpointer user_data);
 
 // slider:
 GtkWidget *dt_bauhaus_slider_new(dt_bauhaus_t *bh, dt_gui_module_t *self);
